Subindo Detalhes de Categoria finalizado

+++ app/Http/Controllers/Admin/PlanController.php
@@ -53,10 +53,18 @@ class PlanController extends Controller
 
     public function destroy($url)
     {
-        $plan = $this->repository->where('url', $url)->first();
+        $plan = $this->repository
+                        ->with('details')
+                        ->where('url', $url)
+                        ->first();
 
         if(!$plan){
             return redirect()->back();
+        }
+
+        if($plan->details->count() > 0) {
+            return redirect()->back()
+                                ->with('error', 'Existem detalhes vinculado a esse plano, não pode ser excluido a menos que exclua os detalhes.');
         }
 
         $plan->delete();
